pyproject.toml
[project]
name = "agent-project"
version = "0.1.0"
description = "A Python project using MCP and FastAPI with Google Sheets integration"
authors = [
    {name = "Riyanshi Bohra", email = "riyanshibohraa@gmail.com"}
]
dependencies = [
    "mcp>=0.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "pandas>=2.0.0",
    "gspread>=5.12.0",
    "google-auth>=2.23.0",
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
    "asyncio>=3.4.3",
]
requires-python = ">=3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[project.scripts]
agent-project = "src.server:main" 